fix(daisy): align audition and resident sample gain
- Share the metadata dB law and retain attenuation at minus 24 dB.
- Support positive streamed gain with vendored CMSIS saturation and exact unity.
- Test the full control range and retain physical listening and timing gates.

+++ firmware/daisy/src/audio/audio_engine.h
@@ -273,7 +273,7 @@ uint8_t DebugActiveVoices();
 struct StreamDebugState {
     uint16_t sample_id = 0;
     uint32_t start = 0, end = 0, loop_start = 0, loop_end = 0, rewinds = 0;
-    int16_t gain_q15 = 0;
+    int32_t gain_q15 = 0;
     bool open = false, loop = false;
 };
 StreamDebugState DebugStreamState();
